test: UserWordService 테스트 추가 - 사용자 선호 단어 저장 및 조회 기능 검증
